Refactored "file reference" and "line number". The file reference is not needed, and misleading - instead, just pick up the eResource using EMF - this indicates where an object came from.
The line number is misleading - some instances of b3 does not come from source files, and the reference is misleading. Instead, in situations where the parser is used, it is possible to obtain the line and offset information in the parser.

The exceptions that related to an expression now have the expression as a member to facilitate getting the line and offset information from the parser.