Code cleanup using Eclipse 4.4 IDE
Some improvements were made to the code formatter in Eclipse 4.4
that leads to lots of annoying changes when auto formatting the code.
One example is that the formatter now removes White space in javadoc
comments.

This commit changes no code. It's just the result of auto formatting
using the new inproved formatter in a 4.4 IDE.